redisearch/openbeerdb/frontend/requirements.txt
redis==3.0.1
redisearch==0.7.1
python-dotenv==0.10.3
Flask==1.0.2
Flask-WTF==0.14.2
Werkzeug==0.15.3
six==1.12.0
Jinja2==2.10.1
WTForms==2.2.1
